feat(composite): wire CompositeTab into sub-tab structure
- CompositeTab is now a tab container with Plans/Wealth Journey/Stress Test/Year-by-Year
- Existing phase config + scenario comparison move to PlansTab
- Wealth journey chart moves to WealthJourneyTab
- Year-by-year timeline moves to YearByYearTab
- StressTestTab remains placeholder (Task 6.2)
- Each tab reads from CompositeProjectionContext
- Sub-tab nav mirrors PlanTabNavigation visual style for consistency
